Optimize AbstractBytes initialization
Current implementation was calling getClass().getCanonicalName() for each
created AbstractBytes instance. This has a performance impact even though
the class name is only needed for the toString() method implementation.
toString() is usually needed only for debugging so this changes it so that
the performance impact is hit only when toString() is called.

Also TO_STRING_SIZE_LIMIT was made static (saves memory) and Gradle Wrapper
was upgraded to version 3.3.